Integrated the new projection method.
Let a point P(x,y,z) be a point in the world. To project it onto the x-y
plane, consider the line L connecting P and Q(0,0,r) where r < 0. Find the
point of intersection of L with the x-y plane. This is the projected
point.

The parameter r is fixed, and currently set to -800. Values too high or
too low make skew the field of view uncomfortably. The current value
seems to work nicely.